Update table generation code towards being able to make custom tables.
Info for cutom tables is in Custom.tables.dat.  Added read_custom.pl 
to parse and verify this file.

copy_tables adds a --custom option to process the custom tables.

Currently the only type of custom table implemented is for articles
rated by a particular project that are also in a particular category.